\name{AutoSULT-internal}
\Rdversion{2.10}
\title{Internal AutoSULT Objects}
\alias{n.defann}
\alias{n.deftermanndue}
\alias{n.ktqx}
\alias{n.nEx}
\alias{n.termanndue}
\alias{n.termins}
\alias{n.tpx}
\alias{n.tqx}
\alias{n.WLannUDD}
\alias{n.WLannW2}
\alias{s.anncertdue}
\alias{s.annguardue}
\alias{s.conttermann}
\alias{s.contWLann}
\alias{s.defconttermann}
\alias{s.defcontWLann}
\alias{s.defguartermanndue}
\alias{s.defmtermann}
\alias{s.defmWLann}
\alias{s.deftermanndue}
\alias{s.deftermins}
\alias{s.defterminsc}
\alias{s.defterminsm}
\alias{s.defwl}
\alias{s.defWLanndue}
\alias{s.defwlc}
\alias{s.defwlm}
\alias{s.ellx}
\alias{s.exp}
\alias{s.frac}
\alias{s.insend}
\alias{s.insendc}
\alias{s.insendm}
\alias{s.ktqx}
\alias{s.minus}
\alias{s.mtermann}
\alias{s.mWLann}
\alias{s.nEx}
\alias{s.plus}
\alias{s.termanndue}
\alias{s.termins}
\alias{s.terminsc}
\alias{s.terminsm}
\alias{s.times}
\alias{s.tpx}
\alias{s.tqx}
\alias{s.WLanndue}
\alias{s.npxy}
\alias{s.nqxy}
\alias{s.npx2}
\alias{s.nqx2}
\alias{s.nExy}
\alias{s.nExy2}
\alias{s.terminsj2}
\alias{s.WLjt}
\alias{s.WLjt2}
\alias{s.termannduejt}
\alias{s.annduerev}
\alias{s.WLannduejt}
\description{
  Internal AutoSULT functions.  Functions s.xxx() create commonly-used strings for typesetting, and functions n.xxx() perform commonly-used calculations.
}
\details{
  These are not intended to be called by the user and
  are only used by other functions.
}
%\seealso{}
\keyword{internal}
