\name{SSD}

\docType{data}

\alias{SSD}
\title{Standard Sickness-Death Table.}

\description{This is the Standard Sickness-Death Model as defined in Figure 8.4 (pg. 290) and Appendix D.3 of Dickson et al.  Death transition intensities are based on the Makeham mortality model.  All calculations are done an interest rate of 5%.
    }

\usage{SSD}

\format{See above.}

\source{Dickson et al., 3rd ed., pg. 280, 744--746}

\references{Dickson, D.C.M., Hardy, M.R., Waters, H.R., 2020 \emph{Actuarial Mathematics for Life Contingent Risks}, Third Edition, Cambridge, UK: Cambridge University Press.
}

\keyword{datasets}

