\name{SULT}

\docType{data}

\alias{SULT}
\alias{SULTjt}
\title{Standard Ultimate Life Table.}

\description{This is the Standard Ultimate Life Table (SULT) a.k.a. Standard Ultimate Survival Model (SUSM) as defined on pg. 82 of Dickson et al.  It is based on the Makeham mortality model.  The object SULT refers to the single life table and SULTjt refers to the joint life table, where the two lives are independent.  All calculations are done an interest rate of 5%.
    }

\usage{SULT}

\format{See above.}

\source{Dickson et al., 3rd ed., pg. 82}

\references{Dickson, D.C.M., Hardy, M.R., Waters, H.R., 2020 \emph{Actuarial Mathematics for Life Contingent Risks}, Third Edition, Cambridge, UK: Cambridge University Press.
}

\keyword{datasets}

