\name{SULT}

\docType{data}

\alias{SULT}
\title{Standard Ultimate Life Table.}

\description{This is the single life Standard Ultimate Life Table (SULT) a.k.a. Standard Ultimate Mortality Model (SUMM) as defined on pg. 65 of Dickson et al.  It is based on the Makeham mortality model.
    }

\usage{SULT}

\format{See above.}

\source{Dickson et al., 2nd ed., pg. 65}

\references{Dickson, D.C.M., Hardy, M.R., Waters, H.R., 2013 \emph{Actuarial Mathematics for Life Contingent Risks}, Second Edition, Cambridge, UK: Cambridge University Press.
}

\keyword{datasets}

